requirements.txt
google-genai>=1.0.0
pywin32>=306
python-dotenv>=1.0.0
pyinstaller>=6.0.0
